Update README.md
The series and movies here, for me, is the best of all time. Most of the movies included here has a lot of part which will caught your attention and be intrigued to the next part of the movie. On the other hand, most of the series that i've inserted here has a lot of humor which will keep the watchers to be entertained.